Updated migration script to create a checksum table.
